refactor: 개선된 대기 예약 조회 및 순위 계산 로직

+++ src/main/java/roomescape/reservation/dto/ReservationWithRank.java
@@ -2,5 +2,5 @@ package roomescape.reservation.dto;
 
 import roomescape.reservation.domain.Reservation;
 
-public record ReservationWithRank(Reservation waiting, Long rank) {
+public record ReservationWithRank(Reservation waiting, int rank) {
 }
